Refactor Assets and Results

+++ backend/src/main/java/com/ares/backend/model/Bonds.java
@@ -4,7 +4,7 @@ import com.ares.backend.AssetType;
 
 public class Bonds extends Asset {
 
-    public Bonds(float startcapital, int years) {
+    public Bonds(double startcapital, int years) {
         super(AssetType.BONDS, startcapital, years, 0.03f, 0.05f);
     }
 }
