+++ backend/src/main/java/com/ares/backend/model/RawMaterials.java
@@ -4,7 +4,7 @@ import com.ares.backend.AssetType;
 
 public class RawMaterials extends Asset {
 
-    public RawMaterials(float startcapital, int years) {
+    public RawMaterials(double startcapital, int years) {
         super(AssetType.RAW_MATERIALS, startcapital, years, 0.06f, 0.10f);
     }
 }
