+++ backend/src/main/java/com/ares/backend/model/RealEstates.java
@@ -4,7 +4,7 @@ import com.ares.backend.AssetType;
 
 public class RealEstates extends Asset {
 
-    public RealEstates(float startcapital, int years) {
+    public RealEstates(double startcapital, int years) {
         super(AssetType.REAL_ESTATES, startcapital, years, 0.05f, 0.07f);
     }
 }
