+++ backend/src/main/java/com/ares/backend/model/Stocks.java
@@ -4,7 +4,7 @@ import com.ares.backend.AssetType;
 
 public class Stocks extends Asset {
 
-    public Stocks(float startcapital, int years) {
+    public Stocks(double startcapital, int years) {
         super(AssetType.STOCKS, startcapital, years, 0.08f, 0.15f);
     }
 }
